requirements.txt
Django==1.4.3
Fabric==1.7.0
PyYAML==3.10
boto==2.8.0
django-autoslug==1.6.1
distribute==0.6.35
django-debug-toolbar==0.9.4
nltk==2.0.4
numpy==1.7.0
paramiko==1.10.1
psycopg2==2.4.6
pycrypto==2.6
python-dateutil==2.1
pytz==2012j
six==1.3.0
stevedore==0.8
tweepy==2.0
virtualenv==1.8.4
virtualenv-clone==0.2.4
virtualenvwrapper==3.6
wsgiref==0.1.2
